+++ test/requirements.txt
@@ -1,3 +1,2 @@
 addict
 modelscope
-git+https://github.com/deepseek-ai/DeepSeek-VL2.git
